// src/app/model/datagaji.model.ts
import { ProfileModel } from './profile.model';
export class DataGajiModel {
    id: string;
    gapok: string;
    gajihonor: string;
    tjistri: string;
    tjanak: string;
    tjjabatan: string;
    tjkesehatan: string;
    tjkhusus: string;
    tjtransport: string;
    tjperalihan: string;
    tjpangan: string;
    tjpengabdian: string;
    tht: string;
    gajikotor: string;
    pttht: string;
    ptpajak: string;
    gabersih: string;
    pembulatan: string;
    tglgaji: string;
    dataPegawai: ProfileModel;
    kol: string;
    createat: string;
    updateat: string;
}