// src/app/model/kolomtunjangan.model.ts
import { ParameterJenisTunjanganKaryawan } from './parameterjenistunjangan.model';
export class KolomTunjangan {
  id: string;
  kolom: string;
  kantor: string;
  createat: string;
  updateat: string;
  jenisTunjangan: ParameterJenisTunjanganKaryawan;
  nominal: string;
  jabatan: string;
  status: string;
  persen: string;


}